Add folder group share selected state

+++ app/src/main/java/mega/privacy/android/app/lollipop/adapters/ShareContactsHeaderAdapter.java
@@ -6,6 +6,8 @@ import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import androidx.core.content.ContextCompat;
 import androidx.recyclerview.widget.RecyclerView;
+
+import android.graphics.Color;
 import android.util.DisplayMetrics;
 import android.util.TypedValue;
 import android.view.Display;
@@ -217,28 +219,31 @@ public class ShareContactsHeaderAdapter extends RecyclerView.Adapter<ShareContac
                 holder.avatar.setImageBitmap(getAvatarShareContact(mContext, contact));
                 UserAvatarListenerShare listener = new UserAvatarListenerShare(mContext, holder);
 
-                File avatar = buildAvatarFile(mContext,mail + ".jpg");
-                Bitmap bitmap = null;
-                if (isFileAvailable(avatar)){
-                    if (avatar.length() > 0){
-                        BitmapFactory.Options bOpts = new BitmapFactory.Options();
-                        bOpts.inPurgeable = true;
-                        bOpts.inInputShareable = true;
-                        bitmap = BitmapFactory.decodeFile(avatar.getAbsolutePath(), bOpts);
-                        if (bitmap == null) {
-                            avatar.delete();
-                            megaApi.getUserAvatar(contact.getMegaContactAdapter().getMegaUser(),buildAvatarFile(mContext,mail + ".jpg").getAbsolutePath(),listener);
+                if (contact.getMegaContactAdapter().isSelected()) {
+                    holder.itemLayout.setBackgroundColor(ContextCompat.getColor(mContext, R.color.new_multiselect_color));
+                    holder.avatar.setImageResource(R.drawable.ic_select_folder);
+                } else {
+                    holder.itemLayout.setBackgroundColor(Color.WHITE);
+                    File avatar = buildAvatarFile(mContext, mail + ".jpg");
+                    Bitmap bitmap = null;
+                    if (isFileAvailable(avatar)) {
+                        if (avatar.length() > 0) {
+                            BitmapFactory.Options bOpts = new BitmapFactory.Options();
+                            bOpts.inPurgeable = true;
+                            bOpts.inInputShareable = true;
+                            bitmap = BitmapFactory.decodeFile(avatar.getAbsolutePath(), bOpts);
+                            if (bitmap == null) {
+                                avatar.delete();
+                                megaApi.getUserAvatar(contact.getMegaContactAdapter().getMegaUser(), buildAvatarFile(mContext, mail + ".jpg").getAbsolutePath(), listener);
+                            } else {
+                                holder.avatar.setImageBitmap(bitmap);
+                            }
+                        } else {
+                            megaApi.getUserAvatar(contact.getMegaContactAdapter().getMegaUser(), buildAvatarFile(mContext, mail + ".jpg").getAbsolutePath(), listener);
                         }
-                        else{
-                            holder.avatar.setImageBitmap(bitmap);
-                        }
+                    } else {
+                        megaApi.getUserAvatar(contact.getMegaContactAdapter().getMegaUser(), buildAvatarFile(mContext, mail + ".jpg").getAbsolutePath(), listener);
                     }
-                    else{
-                        megaApi.getUserAvatar(contact.getMegaContactAdapter().getMegaUser(),buildAvatarFile(mContext,mail + ".jpg").getAbsolutePath(),listener);
-                    }
-                }
-                else{
-                    megaApi.getUserAvatar(contact.getMegaContactAdapter().getMegaUser(),buildAvatarFile(mContext,mail + ".jpg").getAbsolutePath(),listener);
                 }
             }
         }
